Added AI classifier to main.py and integrated the pipeline through ai_detector.py
After, BM25 and SBERT, we get top K articles and contents. The AI model performs paragraph wise segmenting on the contents of the article. The threshold of the trained model was 0.82 that means if probability> 0.82, each paragraph will be independently flagged AI generated. For the overall probabilities, it aggregates the probability of all paragraphs, and for the percentage, it's (AI flagged Paragraphs/Overall Paragraphs)